Did set up for Play Action, could potentially work. Internal track is active
