+++ app/google-services.json
@@ -10,7 +10,7 @@
       "client_info": {
         "mobilesdk_app_id": "1:1093901500458:android:0284db82b85375f0918353",
         "android_client_info": {
-          "package_name": "com.example.accessiblemesseging"
+          "package_name": "com.mana.accessiblemesseging"
         }
       },
       "oauth_client": [
